ssup changes

+++ product/models.py
@@ -37,13 +37,16 @@ class Product(models.Model):
             )
         ]
 
+
 class Promotion(models.Model):
-    product = models.OneToOneField(Product, on_delete=models.CASCADE, related_name='promotion')
+    product = models.OneToOneField(
+        Product, on_delete=models.CASCADE, related_name='promotion')
     initial_price = models.DecimalField(max_digits=10, decimal_places=2)
     start_date = models.DateField()
     end_date = models.DateField()
-    current_price = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
-    discount_percentage = models.DecimalField(max_digits=5, decimal_places=2, blank=True, null=True)
+    current_price = models.DecimalField(
+        max_digits=10, decimal_places=2, blank=True, null=True)
+    discount_percentage = models.IntegerField(blank=True, null=True)
 
     def __str__(self):
         return f"Promotion for {self.product.title}"
